+++ src/index.css
@@ -13,7 +13,6 @@ ol, ul {
   margin-left: auto;
   margin-right: auto;*/
   margin: 0 auto;
-  width: 340px;
   border: 50px solid #542;
 }
